@@ -46,7 +46,7 @@ namespace Moq.AutoMock.Tests
         public void It_chooses_the_ctor_with_arguments()
         {
             var ctor = selector.SelectFor(typeof(WithDefaultAndSingleParameter), new Type[0], DefaultBindingFlags);
-            Assert.Equal(1, ctor.GetParameters().Length);
+            Assert.Single(ctor.GetParameters());
         }
 
         [Fact]
@@ -60,21 +60,21 @@ namespace Moq.AutoMock.Tests
         public void It_chooses_the_ctor_with_the_most_arguments_when_arguments_are_arrays()
         {
             var ctor = selector.SelectFor(typeof(WithArrayParameter), new Type[0], DefaultBindingFlags);
-            Assert.Equal(1, ctor.GetParameters().Length);
+            Assert.Single(ctor.GetParameters());
         }
 
         [Fact]
         public void It_wont_select_if_an_argument_is_sealed_and_not_array()
         {
             var ctor = selector.SelectFor(typeof(WithSealedParameter), new Type[0], DefaultBindingFlags);
-            Assert.Equal(0, ctor.GetParameters().Length);
+            Assert.Empty(ctor.GetParameters());
         }
 
         [Fact]
         public void It_will_select_if_an_argument_is_sealed_and_supplied()
         {
             var ctor = selector.SelectFor(typeof(WithSealedParameter), new Type[] { typeof(string) }, DefaultBindingFlags);
-            Assert.Equal(1, ctor.GetParameters().Length);
+            Assert.Single(ctor.GetParameters());
         }
 
         [Fact]
